+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.0.0 / 2019-01-07
+==================
+* [FEATURE] Converted the options object into a class.
+* [FEATURE] Converted the sizes option in options to its own set of interfaces.
+* [MISC] Updated all dependencies to their latest versions.
+* [MISC] Changed license year to reflect new year.
+* [MISC] Tested to make sure it still works with the latest major version of linkquest.
+
 0.1.1 / 2019-11-11
 ==================
 * [HOTFIX] Updated dependencies with security vulnerabilities.